Move handle submit logic into the Toolbox Steps that use it
